фиксы замечаний и не уверен в правильности обобщения
