+++ ReminderClassLibrary/ReminderClassLibrary.csproj
@@ -93,7 +93,8 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RemindEventArgs.cs" />
     <Compile Include="RemindTask.cs" />
-    <Compile Include="RemindRepository.cs" />
+    <Compile Include="RemindFileRepository.cs" />
+    <Compile Include="TaskScheduler.cs" />
     <Compile Include="TaskStatus.cs" />
   </ItemGroup>
   <ItemGroup>
